making it check for interactive

+++ R/md_helpers.R
@@ -53,10 +53,13 @@ bibliography: bibliography_file.json', sep = "\n")
 #' @return string
 ft_remove_figures = function()
 {
-  cat("doc_type = knitr::opts_knit$get('rmarkdown.pandoc.to')
-if (doc_type %in% 'docx') {
-  unlink(here::here('path', 'to', 'documentname_files'), recursive = TRUE)
-}", sep = "\n")
+  cat("
+  if (!interactive()) {
+    doc_type = knitr::opts_knit$get('rmarkdown.pandoc.to')
+    if (doc_type %in% 'docx') {
+      unlink(here::here('path', 'to', 'documentname_files'), recursive = TRUE)
+    }
+  }", sep = "\n")
 }
 
 #' plot list includes
